+++ hspec-discover/integration-test/with-config/SpecConfig.hs
@@ -1,0 +1,9 @@
+module SpecConfig where
+
+import Test.Hspec.Runner
+import Test.Hspec.Core.Formatters.V2
+
+config :: Config -> IO Config
+config c = return c {
+  configFormat = Just $ formatterToFormat progress
+}