+++ CHANGES.markdown
@@ -1,6 +1,7 @@
 ## Changes in next
   - Better support for `--color=auto` on Windows
   - Add support for spec hooks to `hspec-discover` (via `SpecHooks.hs`)
+  - Add support for custom configs to `hspec-discover` (via `SpecConfig.hs`)
   - Propagate exceptions from `Test.Hspec.Core.Format.monadic`
 
 ## Changes in 2.8.2